+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         android:theme="@style/Theme.DiaryByCompose"
         tools:targetApi="31" >
         <activity
-            android:name="com.jm.diarybycompose.MainActivity"
+            android:name="com.jm.diarybycompose.ui.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.DiaryByCompose" >
