+++ app/src/main/java/com/jm/diarybycompose/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.jm.diarybycompose.compose.theme
+package com.jm.diarybycompose.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
